fiing timer, will add pause button
